Use C89 compatible comments
These will cause clang -std=c89 to fail on recent versions of the
Macintosh, and fit the style of the rest of the code better anyway.

As to why these assets are commented out is not well explained, given
that as an Assert statement one would naively assume they have no
affect on production-built code.

Signed-off-by: Daniel Farina <drfarina@acm.org>