Header cleanup: move <stdarg.h> to <base.h>
Managing stdarg.h has caused portability problems (#11).

But when including in one header, it will spread to others
so it's still confusing when it is available or not.

Also it seems on Linux it already is included by some
system header, because va_list works.

It's simpler to tag is as base header that is always available.

This commit also removes few headers that are already in base.