tls: don't use SSL_VERIFY_FAIL_IF_NO_PEER_CERT
The problem is that tls_config does not know whether it's
for client or server so CA is always configured.  So
client without certs will be rejected.

This should be OK as server can always check if cert exists
with tls_get_peer_cert().  It makes hard to have server
that just wants validated cert.  But this seems rather
uncommon scenario.