Fix TLSv1.3 support with certain OpenSSL versions
With OpenSSL 1.1.1 and 1.1.1a (but not before or after), a PgBouncer
to backend server connection attempt will fail with

    WARNING tls_sbufio_recv: unexpected handshake, closing connection

if TLSv1.3 is being used (which is the case automatically, unless
disabled via openssl.cnf).

This is because the info callback is being called for TLSv1.3 by
OpenSSL in a way that the code was not expecting.  See
<https://www.openssl.org/news/changelog.html#x1> for details.

The info callback was trying to prevent renegotiation, but that is no
longer applicable with TLSv1.3, so skip that if we are using TLSv1.3.

researched by Francesco Canovai