Update ssl test certificates and keys
Newer Debian and Ubuntu versions require that RSA and DHE keys are at
least 2048 bit long.  This is currently causing the ssl tests to fail
there because the test certificates and keys have been created with
only 1024 bits.

Update the parameters to create the test files and create a new set of
test files.