Use systemd journal for logging when in use
When we detect that stderr is going to the systemd journal, we use the
function sd_journal_print() instead of fprintf(stderr, ...) for log
output.  Primarily, this avoids printing duplicate timestamp and pid,
thus making the log a bit cleaner.  Secondarily, this adds metadata
such as the severity to the logs, so that if the journal gets sent on
to syslog, the messages have useful metadata attached.