Ignore certain GCC warnings in SHList test
New versions of GCC warn about pointer magic that SHList is actively
exploiting. This ignores those warnings.

Obviously it would be better to fix the warnings, but afaict that would
require huge changes to SHList. Since SHList is not used by PgBouncer
the impact of this is quite small and its not worth the effort to fix
this at the moment. This change at least makes CI green again.