antimake: split LangObjTarget rule
GNU make 4.4 started to show warning:

	warning: pattern recipe did not update peer target '...'

That is because make expects rule in form

	%.o %.lo: %.c

build both targets in one execution, which does happen when
building .lo objects, but not when building .o objects.

But this dependency between .o<>.lo is unnecessary complexity,
as no rule uses both .o and .lo of same object file as input.

Thus is is enough to duplicate the rule for both cases
and let the target pick which one it wants to depend on.