Fix for https://github.com/pgbouncer/pgbouncer/issues/713 (#53)
This is a fix for:
https://github.com/pgbouncer/pgbouncer/issues/713

The issue was pretty simple: the CSR was lacking a filled DN section. 
This was tracked down by reverse-engineering the communication done by
Postgres (which works with DataStudio/Looker without any issues) and
Postgres via pgbouncer.
After comparing packets sent on the handshake to DataStudio/Looker by
those two, I was able to provide a small fix resolving the linked issue.